backend/requirements.txt
fastapi>=0.111,<1
uvicorn[standard]>=0.30,<1
pydantic-settings>=2.3,<3
SQLAlchemy>=2.0,<3
alembic>=1.13,<2
psycopg[binary]>=3.1,<4
PyJWT>=2.8,<3
passlib[bcrypt]>=1.7.4,<2
dashscope>=1.14,<2
httpx>=0.27,<1
redis>=5.0,<7
celery>=5.3,<6
requests>=2.31,<3
oss2>=2.18,<3
numpy>=1.26,<3
scikit-learn>=1.4,<2
hdbscan>=0.8.33,<1
sentence-transformers>=3.0,<4
pytest>=8.0,<9

black>=24.4,<25
isort>=5.13,<6
mypy>=1.10,<2
